DEV - Claude Code - Agents, Memory + Hooks/agents/requirements.txt
# Multi-Agent System Requirements

# Core Dependencies
aiohttp>=3.8.0
python-dotenv>=0.19.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-asyncio>=0.21.0

# Optional Performance
# uvloop>=0.17.0  # Faster async event loop (Unix only)

# Development Tools
black>=22.0.0  # Code formatting
flake8>=4.0.0  # Linting
mypy>=0.991    # Type checking

# Monitoring (optional)
# prometheus-client>=0.15.0  # Metrics
# psutil>=5.9.0  # System monitoring